deleted 2 links in the 'R in organization' section
Since we have a comprehensive list of conferences in 2017 in the 'Resources' as well as we have mentioned some of them in the 'upcoming events'. I feel we don't have to repeatedly mention the conferences again here in 'R in organization'.

It will look redundant and make readers feel that we don't have anything else to post. or just to burden our readers by reading the repetitive posts.