The impact of COVID-19 on healthcare
Dear rweekly team,
I'd like to add an open source paper dealing with the impact of the COVID-19 pandemic on neurosurgical outpatient numbers and elective procedures at a German university hospital to the section "R in the Real World". Please find the link to the article in the newly generated fork. The underlying pseudonymized data and R code will be available upon request from me (senior author of the paper, BV). Suggested figure: Figure 1b of the paper.

Kind regards,

Ben Voellger