Update draft.md (#1509)
Added a link the slides for the presentation i recently did for the "CopenhagenR – useR Group". 

I was a little in doubt if I should link directly to the slides instead, but chose to link to the page on my website that holds the link in order to add some context to the presentation – especially the link to the user group so people can join up if they want to.